// code2/src/app/modules/admin/programs/program-dashboard/program.types.ts
export interface Program {
    programId: number;
    managerName: string;
    title: string;
    colorCode: string;
    startDate: string;
    endDate: string;
    description: string;
    shortDesc: string;
    status: string;
    rating: number;
    coachCount:number;
    coacheeCount: number;
    sessionCount: number;
    engagementCount: number;
    isEditable: boolean;
    countDays: number;
    programTitle: string;
}